goprotobuf: Split encoding of repeated int32 and repeated uint32 fields.
This is a follow-on from a previous change where we split the singular int32/uint32.
It was an oversight not to fix the repeated versions at the same time.

Also simplify setEncAndDec for repeated fields while I'm here.
We were doing a lot of unnecessary work: unnecessarily doing two layers of
switches, plus trying to handle a bunch of field types we never generate.

Fixes issue 52.

LGTM=djd
R=djd
CC=golang-codereviews
https://codereview.appspot.com/123240043